Implements slot scheduling for customers
Adds functionality for customers to schedule available slots.

Introduces a new endpoint for retrieving available slots based on professional and service.
Adds an endpoint for scheduling a slot by assigning a customer.
Adds exception handling for cases where the slot is not found, not available or the customer is invalid.
Creates DTOs for scheduling and displaying slots.
Exposes endpoint to retrieve all customers.